// src/types/Datasworn.d.ts
export type AssetId = string
// ...
export type AtlasEntryId = string
// ...
export type DelveSiteDomainId = string

/**
 * A unique ID for a DelveSite.
 * @pattern ```javascript
 * /^([a-z0-9_]{3,})\/delve_sites\/([a-z][a-z_]*)$/
 * ```
 * @example "delve/delve_sites/alvas_rest"
 */
export type DelveSiteId = string

/**
 * A unique ID for a DelveSiteTheme.
 * @pattern ```javascript
 * /^([a-z0-9_]{3,})\/site_themes\/([a-z][a-z_]*)$/
 * ```
 * @example "delve/site_themes/hallowed"
 */
export type DelveSiteThemeId = string

/**
 * A `snake_case` key used in a Datasworn dictionary object.
 * @remarks If you need to generate a key from a user-provided label, it's recommended to use a 'slugify' function/library, e.g. https://www.npmjs.com/package/slugify for NodeJS.
 * @pattern ```javascript
 * /^([a-z][a-z_]*)$/
 * ```
 */
export type DictKey = string
// ...
export type MoveId = string
// ...
export type NpcId = string
// ...
export type OracleCollectionId = string

/**
 * A unique ID for an OracleTable.
 * @pattern ```javascript
 * /^([a-z0-9_]{3,})\/oracles(\/([a-z][a-z_]*)){1,3}\/([a-z][a-z_]*)$/
 * ```
 * @example "starforged/oracles/core/action"
 * @example "starforged/oracles/character/names/given"
 * @example "starforged/oracles/planets/furnace/settlements/terminus"
 */
export type OracleTableId = string
// ...
export type OracleTableRowId = string

/**
 * A unique ID for a Rarity.
 * @pattern ```javascript
 * /^([a-z0-9_]{3,})\/rarities\/([a-z][a-z_]*)$/
 * ```
 * @example "classic/rarities/ayethins_journal"
 */
export type RarityId = string
// ...
export type CssColor = string

/**
 * @pattern ```javascript
 * /^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$/
 * ```
 */
export type SemanticVersion = string

/**
 * Metadata describing the original source of this item
 */
export interface Source {
	/**
	 * The title of the source document.
	 * @example "Ironsworn Rulebook"
	 * @example "Ironsworn Assets Master Set"
	 * @example "Ironsworn: Delve"
	 * @example "Ironsworn: Starforged Rulebook"
	 * @example "Ironsworn: Starforged Assets"
	 * @example "Sundered Isles"
	 */
	title: string
	/**
	 * The page number where this item is described in full.
	 */
	page?: number
	authors: Array<{
		/**
		 * @example "Shawn Tomkin"
		 */
		name: string
		/**
		 * An optional email contact for the author
		 */
		email?: string
		/**
		 * An optional URL for the author's website.
		 */
		url?: string
	}>
	/**
	 * The date of the source documents's last update, formatted YYYY-MM-DD. Required because it's used to determine whether the data needs updating.
	 * @remarks You may prefer to deserialize this as a Date object.
	 */
	date: string
	/**
	 * An absolute URL where the source document is available.
	 * @example "https://ironswornrpg.com"
	 */
	url: string
	/**
	 * An absolute URL pointing to the location where this element's license can be found.
	 *
	 * A `null` here indicates that the content provides __no__ license, and is not intended for redistribution.  Datasworn's build process skips unlicensed content by default.
	 * @example "https://creativecommons.org/licenses/by/4.0"
	 * @example "https://creativecommons.org/licenses/by-nc-sa/4.0"
	 */
	license: string | null
}

export interface Suggestions {
	oracles?: OracleTableId[]
	assets?: AssetId[]
	moves?: MoveId[]
	site_domains?: DelveSiteDomainId[]
	site_themes?: DelveSiteThemeId[]
	npcs?: NpcId[]
	atlas?: AtlasEntryId[]
	rarities?: RarityId[]
}

/**
 * A relative URL pointing to a vector image in the SVG format.
 * @pattern ```javascript
 * /\.svg$/
 * ```
 */
export type SvgImageUrl = string

/**
 * A relative URL pointing to a raster image in the WEBP format.
 * @pattern ```javascript
 * /\.webp$/
 * ```
 */
export type WebpImageUrl = string

export interface I18nHint {
	part_of_speech?: PartOfSpeech
}

/**
 * Internationalization/localization hints for the text content of this object.
 */
export interface I18nHints {
	result?: I18nHint
	summary?: I18nHint
	description?: I18nHint
	template?: {
		result?: I18nHint
		summary?: I18nHint
		description?: I18nHint
	}
}

/**
 * A localized label for an input. In some contexts it may be undesirable to render this text, but it should always be exposed to assistive technology (e.g. with `aria-label` in HTML).
 * @pattern ```javascript
 * /^[^A-Z]+$/
 * ```
 * @i18n
 */
export type InputLabel = string

/**
 * A localized plain text name or label.
 * @i18n
 */
export type Label = string

/**
 * Localized text, formatted in Markdown.
 *
 * It uses some custom syntax; e.g. `{{table:some_oracle_table_id}}` indicates that the referenced oracle table is rendered there part of the source material.
 * @i18n
 */
export type MarkdownString = string
// ...
export type TemplateString = string
// ...
export type DiceExpression = string
// ...
export interface OracleRollTemplate {
	/**
	 * A string template that may be used in place of OracleTableRow#result.
	 * @example "{{result:starforged/oracles/factions/affiliation}} of the {{result:starforged/oracles/factions/legacy}} {{result:starforged/oracles/factions/identity}}"
	 */
	result?: TemplateString
	/**
	 * A string template that may be used in place of OracleTableRow#summary.
	 */
	summary?: TemplateString
	/**
	 * A string template that may be used in place of OracleTableRow#description.
	 */
	description?: TemplateString
}

export interface OracleTableMatchBehavior {
	text: MarkdownString
}

export interface OracleTableRoll {
	/**
	 * The ID of the oracle table to be rolled. A `null` value indicates that it's a roll on the same table.
	 * @default null
	 */
	oracle: OracleTableId | null
	/**
	 * Both Ironsworn and Starforged explicitly recommend *against* rolling all details at once. That said, some oracle results only provide useful information once a secondary roll occurs, such as "Action + Theme".
	 * @default false
	 */
	auto: boolean
	/**
	 * The number of times to roll.
	 * @default 1
	 */
	times: number
	/**
	 * Special rules on how the oracle table roll is performed.
	 * @default "no_duplicates"
	 */
	method: OracleTableRollMethod
	/**
	 * The dice roll to make on the oracle table. Set it to `null` if you just want the table's default.
	 * @default null
	 */
	dice: DiceExpression | null
}

/**
 * Special roll instructions to use when rolling multiple times on a single oracle table.
 *
 *   - `no_duplicates`: Duplicates should be re-rolled.
 *   - `keep_duplicates`: Duplicates should be kept.
 *   - `make_it_worse`: Duplicates should be kept, and they compound to make things worse.
 */
export type OracleTableRollMethod =
	| 'no_duplicates'
	| 'keep_duplicates'
	| 'make_it_worse'

export interface OracleCollection {
	/**
	 * @default
	 * ```javascript
	 * 	{
	 * 		style: "tables"
	 * 	}
	 * ```
	 */
	rendering?: OracleCollectionRendering
	/**
	 * The unique Datasworn ID for this item.
	 */
	id: OracleCollectionId
	/**
	 * The primary name/label for this item.
	 */
	name: Label
	/**
	 * The name of this item as it appears on the page in the book, if it's different from `name`.
	 */
	canonical_name?: Label
	/**
	 * Attribution for the original source (such as a book or website) of this item, including the author and licensing information.
	 */
	source: Source
	suggestions?: Suggestions
	/**
	 * A thematic color associated with this collection.
	 */
	color?: CssColor
	/**
	 * A brief summary of this collection, no more than a few sentences in length. This is intended for use in application tooltips and similar sorts of hints. Longer text should use the "description" key instead.
	 */
	summary?: MarkdownString
	/**
	 * A longer description of this collection, which might include multiple paragraphs. If it's only a couple sentences, use the `summary` key instead.
	 */
	description?: MarkdownString
	images?: WebpImageUrl[]
	/**
	 * An SVG icon associated with this collection.
	 */
	icon?: SvgImageUrl
	/**
	 * This collection's content enhances the identified collection, rather than being a standalone collection of its own.
	 */
	enhances?: OracleCollectionId
	/**
	 * This collection replaces the identified collection. References to the replaced collection can be considered equivalent to this collection.
	 */
	replaces?: OracleCollectionId
	/**
	 * @remarks Deserialize as a dictionary object.
	 */
	contents?: Record<DictKey, OracleTable>
	/**
	 * @remarks Deserialize as a dictionary object.
	 */
	collections?: Record<DictKey, OracleCollection>
}

/**
 * Describes the presentation of this oracle collection, which might represent a group of separate tables, or a single table with additional columns.
 * @remarks Deserialize as a discriminated union/polymorphic object type, using the `style` property as a discriminator.
 */
export type OracleCollectionRendering =
	| OracleCollectionRenderingTables
	| OracleCollectionRenderingMultiTable

export interface OracleCollectionRenderingMultiTable {
	/**
	 * Presented as a single table, with its OracleTable children rendered as columns.
	 */
	style: 'multi_table'
	/**
	 * @remarks Deserialize as a dictionary object.
	 */
	columns: Record<DictKey, OracleCollectionTableColumn>
}

export interface OracleCollectionRenderingTables {
	/**
	 * Presented as a collection of separate tables.
	 */
	style: 'tables'
}
// ...
export interface OracleTable {
	/**
	 * The unique Datasworn ID for this item.
	 */
	id: OracleTableId
	/**
	 * The primary name/label for this item.
	 */
	name: Label
	/**
	 * The name of this item as it appears on the page in the book, if it's different from `name`.
	 */
	canonical_name?: Label
	/**
	 * Attribution for the original source (such as a book or website) of this item, including the author and licensing information.
	 */
	source: Source
	suggestions?: Suggestions
	/**
	 * The roll used to select a result on this table.
	 * @default "1d100"
	 */
	dice: DiceExpression
	_i18n?: I18nHints
	/**
	 * An icon that represents this table.
	 */
	icon?: SvgImageUrl
	images?: WebpImageUrl[]
	/**
	 * A brief summary of the oracle table's intended usage, no more than a few sentences in length. This is intended for use in application tooltips and similar sorts of hints. Longer text should use the "description" key instead.
	 */
	summary?: MarkdownString
	/**
	 * Indicates that this table replaces the identified table. References to the replaced table can be considered equivalent to this table.
	 */
	replaces?: OracleTableId
	/**
	 * A longer description of the oracle table's intended usage, which might include multiple paragraphs. If it's only a couple sentences, use the `summary` key instead.
	 */
	description?: MarkdownString
	/**
	 * Most oracle tables are insensitive to matches, but a few define special match behavior.
	 */
	match?: OracleTableMatchBehavior
	table: OracleTableRow[]
	/**
	 * Describes how how to render this table, when presenting it as a standalone table.
	 */
	rendering?: OracleTableRendering
}
// ...
export type OracleTableRendering =
	| OracleTableRenderingStandalone
	| OracleTableRenderingColumn
	| OracleTableRenderingEmbedInRow

export interface OracleTableRenderingColumn {
	/**
	 * Render as a single column of a table.
	 */
	style: 'column'
}

export interface OracleTableRenderingEmbedInRow {
	/**
	 * Render as a table, within a row in another table.
	 */
	style: 'embed_in_row'
}

export interface OracleTableRenderingStandalone {
	/**
	 * Render as a standalone table.
	 */
	style: 'standalone'
	/**
	 * @remarks Deserialize as a dictionary object.
	 * @default
	 * ```javascript
	 * 	{
	 * 		roll: {
	 * 			label: "Roll",
	 * 			content_type: "roll"
	 * 		},
	 * 		result: {
	 * 			label: "Result",
	 * 			content_type: "result"
	 * 		}
	 * 	}
	 * ```
	 */
	columns: Record<DictKey, OracleTableColumn>
}

/**
 * Represents a row in an oracle table.
 */
export interface OracleTableRow {
	/**
	 * The unique Datasworn ID for this item.
	 */
	id: OracleTableRowId
	/**
	 * The primary text content of this row.
	 */
	result: MarkdownString
	icon?: SvgImageUrl
	/**
	 * Optional secondary text content for this row. Generally, this is longer than `result`.
	 */
	summary?: MarkdownString
	/**
	 * Optional tertiary text content for this row. Generally, this is longer than both `result` and `summary`.
	 */
	description?: MarkdownString
	/**
	 * Further oracle rolls prompted by this table row.
	 */
	rolls?: OracleTableRoll[]
	suggestions?: Suggestions
	/**
	 * Hints that the identified table should be rendered inside this table row.
	 */
	embed_table?: OracleTableId
	template?: OracleRollTemplate
	i18n?: I18nHints
	/**
	 * Low end of the dice range for this table row. `null` represents an unrollable row, included only for rendering purposes.
	 * @default null
	 */
	min: number | null
	/**
	 * High end of the dice range for this table row. `null` represents an unrollable row, included only for rendering purposes.
	 * @default null
	 */
	max: number | null
}